test(fortnite): split the shop tests out of fortnite_test.go
The shop shares nothing with the stats side beyond the provider that serves
both: a different upstream, no account identity, no rate-limit lane of its own,
and now a cache boundary set by Epic's rotation rather than by a staleness
budget. Holding both in one file was already the weakest cohesion in the
package, and adding the rotation cases pushed it past the gate. The package
already keeps parse, season and sharing concerns in their own files; this
follows them.

No test bodies change.
